fix : 잡다한 코드 오류 수정
